requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.7
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
stripe==13.0.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
